+++ tests/Dummy/DummyAirportCamelDto.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tabuna\Map\Tests\Dummy;
+
+class DummyAirportCamelDto
+{
+    public string $airportCode;
+    public string $cityName;
+}